+++ storage/datastore/nosql/dynamodb/connection.go
@@ -1,0 +1,17 @@
+package dynamodb
+
+//import (
+//	"github.com/aws/aws-sdk-go/aws"
+//	"github.com/aws/aws-sdk-go/aws/session"
+//	"github.com/aws/aws-sdk-go/service/dynamodb"
+//	"github.com/aws/aws-sdk-go/service/dynamodb/dynamodbiface"
+//)
+
+// NewDBConnection Instantiates a new DynamoDB Client Connection
+//func NewDBConnection(awsRegion string) dynamodbiface.DynamoDBAPI {
+//	sess := session.Must(session.NewSession(&aws.Config{
+//		Region: aws.String(awsRegion),
+//	}))
+//	client := dynamodb.New(sess)
+//	return client
+//}